+++ .gitignore
@@ -1,1 +1,1 @@
-backend/user_code/*
+app/user_code
