+++ app/index.html
@@ -3,7 +3,11 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Code Input</title>
+    <title>Code Input with Syntax Highlighting</title>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.65.9/codemirror.min.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.65.9/theme/material.min.css">
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.65.9/codemirror.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.65.9/mode/clike/clike.js"></script>
     <style>
         body {
             margin: 0;
@@ -27,16 +31,13 @@
             max-width: 600px;
         }
 
-        .code-input {
+        .CodeMirror {
             width: 100%;
             height: 200px;
             border: 1px solid #ccc;
             border-radius: 5px;
             font-family: "Courier New", monospace;
             font-size: 16px;
-            padding: 10px;
-            resize: none;
-            outline: none;
         }
 
         .submit-btn {
@@ -58,14 +59,22 @@
 <body>
     <div class="code-container">
         <h2>Enter Your Code</h2>
-        <textarea id="codeInput" class="code-input" placeholder="Write your code here..."></textarea>
+        <textarea id="codeEditor"></textarea>
         <button class="submit-btn" onclick="submitCode()">Submit</button>
     </div>
 
     <script>
+        // Инициализация CodeMirror
+        const editor = CodeMirror.fromTextArea(document.getElementById('codeEditor'), {
+            mode: "text/x-c++src", // Подсветка синтаксиса для C++
+            lineNumbers: true,    // Отображение номеров строк
+            theme: "material",    // Тема редактора
+            matchBrackets: true   // Подсветка парных скобок
+        });
+
         async function submitCode() {
-            const code = document.getElementById('codeInput').value;
-    
+            const code = editor.getValue(); // Получение текста из CodeMirror
+            
             const response = await fetch('/submit', {
                 method: 'POST',
                 headers: {
@@ -73,16 +82,14 @@
                 },
                 body: JSON.stringify({ code }),
             });
-    
+
             if (response.ok) {
                 alert('Code submitted successfully!');
-                // Очистить поле ввода после успешной отправки
-                document.getElementById('codeInput').value = '';
+                editor.setValue(''); // Очистить редактор после успешной отправки
             } else {
                 alert('Error submitting code.');
             }
         }
     </script>
-    
 </body>
 </html>
