--- backend/user_code/user_code.cpp
@@ -1,2 +1,0 @@
-#include <iostream>
-int main(){std::cout<<"22 19"; return 0;}
